Refactored executor to use a lighter weight interface so that every allocation would be controlled. VERY BROKEN. DO NOT PUSH.
