+++ TODO.txt
@@ -1,13 +1,5 @@
 === Up next ===
-    [X] Engine config/bootloader (Binary format: defines controller types, collider types, defaults, etc...)
-        [X] JSON-baking of config
-        [X] Reading the bootloader
-        [X] Initializing script variables created by the boot process
-        [ ] Enumerate controller instances
-        [ ] Defaults
-    [ ] Player-facing configuration (.ini files)
-        [ ] Controller config
-        [ ] Script exposure
+    [ ] Fix the bug where controller axis bindings are mysteriously reset sometime after the bootloader loads
     [ ] Tilemap-Entity depenetration
     [ ] Event dispatching (e.g. collisions)
     [ ] Z-ordered deferred drawing
@@ -42,9 +34,9 @@
         [ ] Textured particles (for sparkles, smoke, etc...)
         [ ] Collision? (maybe environment only)
     [ ] Event system
-    [ ] JSON and binary assets
+    [X] JSON and binary assets
         [X] Building JSON assets into binary assets (external tool)
-    [ ] Resolution independence
+    [X] Resolution independence
     [ ] Looping music
     [ ] Positional sound effects
     [ ] GUI engine
@@ -74,5 +66,5 @@
         [X] Press/Release events
         [X] Binding conflict resolution (via silent auto-unbinding)
         [ ] Input history with sequence matching? (fightstick type stuff)
-        [ ] Configuration
+        [X] Configuration
     [ ] Basic editor (probably in a different programming language)
