+++ src/fileutil.cpp
@@ -1,5 +1,5 @@
 #include "assetmanager.h"
-#include "storage.h"
+#include "arrays.h"
 #include "fileutil.h"
 #include "error.h"
 #include "hitbox.h"
