+++ src/fileutil.h
@@ -1,7 +1,8 @@
 #pragma once
 
 #include <cstdio>
-#include "storage.h"
+#include "mempool.h"
+#include "arrays.h"
 #include "result.h"
 #include "error.h"
 #include "util.h"
