+++ src/input.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "storage.h"
+#include "arrays.h"
 #include "entity.h"
 #include "error.h"
 #include "result.h"
