+++ src/sprite.cpp
@@ -1,5 +1,5 @@
 
-#include "storage.h"
+#include "mempool.h"
 #include "sprite.h"
 #include "error.h"
 #include "fileutil.h"
